Revert "feat(@tinacms/form-builder): Combines FormView and FormBuilder"
