+++ packages/@tinacms/form-builder/src/Form.tsx
@@ -20,7 +20,7 @@ import * as React from 'react'
 import { useState, useContext } from 'react'
 import PropTypes from 'prop-types'
 import { Form } from '@tinacms/forms'
-import { FormLegacy } from './FormLegacy'
+import { FormBuilder } from './form-builder'
 import { Field } from 'react-final-form'
 
 interface RenderProps {
@@ -47,11 +47,11 @@ export function TinaForm({ form, children }: Props) {
 
   return (
     <EditingContext.Provider value={isEditing}>
-      <FormLegacy form={form}>
+      <FormBuilder form={form}>
         {() => {
           return children({ isEditing, setIsEditing })
         }}
-      </FormLegacy>
+      </FormBuilder>
     </EditingContext.Provider>
   )
 }
