+++ packages/@tinacms/react-forms/src/index.ts
@@ -19,5 +19,6 @@ limitations under the License.
 export * from './CreateContentMenu'
 export * from './FormActions'
 export * from './FormPortal'
+export * from './FormView'
 export * from './LoadingDots'
 export * from './ResetForm'
