+++ packages/react-tinacms-inline/src/inline-form.tsx
@@ -18,7 +18,7 @@ limitations under the License.
 
 import * as React from 'react'
 import { FormRenderProps } from 'react-final-form'
-import { FormLegacy, Form, useCMS } from 'tinacms'
+import { FormBuilder, Form, useCMS } from 'tinacms'
 import { Dismissible } from 'react-dismissible'
 import { RBIEPlugin } from './rbie/plugins/rbie-plugin'
 import { InlineFieldsRenderer } from './rbie/components/inline-fields-renderer'
@@ -89,7 +89,7 @@ export function InlineForm({ form, children }: InlineFormProps) {
             setFocussedField('')
           }}
         >
-          <FormLegacy form={form as any}>
+          <FormBuilder form={form as any}>
             {({ form, ...formProps }) => (
               <>
                 {typeof children !== 'function'
@@ -101,7 +101,7 @@ export function InlineForm({ form, children }: InlineFormProps) {
                 {rbie?.active && <InlineFieldsRenderer />}
               </>
             )}
-          </FormLegacy>
+          </FormBuilder>
         </div>
       </Dismissible>
     </InlineFormContext.Provider>
